Peptide Correlation Analysis (PeCorA)

Perform Peptide Correlation Analysis (PeCorA) as described in https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8592057/.

PeCorA(
  m,
  proteins,
  treatment,
  control_group = NULL,
  protein_column = "Protein",
  peptide_column = "Peptide",
  standardize_samples = TRUE,
  center_peptides = TRUE,
  median_mod = TRUE,
  verbose = FALSE
)

Arguments

m

MSnSet object with pData and fData tables. fData must have columns protein_column and peptide_column. The exprs data should be aggregated such that features are unique combinations of peptide and protein.

proteins

character; one or more proteins in fData(m)[[protein_column]] with peptides to test. Only those proteins with at least 2 peptides will be considered.

treatment

character; the name of column in pData(m) containing treatment information. The data can be numeric or categorical. PeCorA will test for an interaction between treatment and all peptides that map to the proteins in proteins.

control_group

character; optional control group. Only applicable if treatment is categorical. See center_peptides parameter below for details.

protein_column

character; the name of a column in fData(m) containing protein information. Default is "Protein".

peptide_column

character; the name of a column in fData(m) containing peptide information. Default is "Peptide".

standardize_samples

logical; whether to standardize samples (columns) in exprs data.

center_peptides

logical; whether to mean-center each peptide. If the data in the treatment column is categorical and control_group is not NULL, peptides will be zero-centered on the control group mean. Otherwise, the mean of all samples are used for centering.

median_mod

logical; whether to use a modified version of PeCorA. See details.

verbose

logical; whether to produce progress messages.

Value

An MSnSet where the fData table includes new or updated columns "pecora_pval_treatment" and "pecora_adj_pval_treatment".

Details

By default, the MSnSet provided by m will be normalized as described in the PeCorA publication: standardize columns and mean-center peptides using the mean of the control_group samples (if provided) or the mean of all samples.

If median_mod = TRUE, the intensities of "all other peptides" will be aggregated to their per-sample medians. This way, all peptides mapping to a given protein (aside from the peptide of interest as the function loops over all peptides) are not treated as 'independent' observations. The effect is that p-values are much higher (less significant) than in the original version of PeCorA.

This function (as well as any helper functions) was based on those from the PeCorA R package (https://github.com/jessegmeyerlab/PeCorA).

References

Dermit, M., Peters-Clarke, T. M., Shishkova, E., & Meyer, J. G. (2021). Peptide Correlation Analysis (PeCorA) Reveals Differential Proteoform Regulation. Journal of proteome research, 20(4), 1972–1980. https://doi.org/10.1021/acs.jproteome.0c00602

See also