Concatenate MSnSet Objects by Features

Combines MSnset objects by concatenating features. Samples must be the same and maintained the same. Pheno data also must be the same.

concatenate_by_features(..., tags = NULL)

Arguments

...

MSnSet objects

tags

vector of characters attached to feature names of corresponding MSnSet objects

Value

MSnSet object with concatenated features

Examples

data(srm_msnset)
dim(msnset)
#> [1] 42 42
featureNames(msnset)
#>  [1] "TIMP3.2"    "TIMP3.1"    "TH.2"       "SLK.1"      "SERPINA3.1"
#>  [6] "SCG5.2"     "SCG5.1"     "S100B.2"    "S100B.1"    "PTPRN.2"   
#> [11] "PREPL.2"    "PREPL.1"    "PLIN3.1"    "PLEC.1"     "PDYN.2"    
#> [16] "PDYN.1"     "PDE2A.2"    "PCLO.1"     "PBXIP1.2"   "PBXIP1.1"  
#> [21] "NEFH.1"     "MYO1D.1"    "MVP.1"      "METAP2.1"   "MBP.3"     
#> [26] "MBP.2"      "MBP.1"      "HP.1"       "HBD.2"      "HBD.1"     
#> [31] "GSTM1.2"    "GFAP.2"     "ERMN.1"     "ENO2.2"     "ENO2.1"    
#> [36] "DKK3.1"     "CRABP1.1p"  "CRABP1.1"   "BSN.2"      "ALDH1A1.3" 
#> [41] "ALDH1A1.1"  "ACTR3B.1"  
msnset3 <- msnset2 <- msnset
# Z-score transform
exprs(msnset2) <- sweep(exprs(msnset2), 1, apply(exprs(msnset2), 1, sd), '/')
# retaining just the sign
exprs(msnset3) <- ifelse(exprs(msnset3) > 0, 1, -1)
m <- concatenate_by_features(msnset, msnset2, msnset3,
                             tags=c("_ori","_zsc","_sgn"))
dim(m)
#> [1] 126  42
featureNames(m)
#>   [1] "TIMP3.2_ori"    "TIMP3.1_ori"    "TH.2_ori"       "SLK.1_ori"     
#>   [5] "SERPINA3.1_ori" "SCG5.2_ori"     "SCG5.1_ori"     "S100B.2_ori"   
#>   [9] "S100B.1_ori"    "PTPRN.2_ori"    "PREPL.2_ori"    "PREPL.1_ori"   
#>  [13] "PLIN3.1_ori"    "PLEC.1_ori"     "PDYN.2_ori"     "PDYN.1_ori"    
#>  [17] "PDE2A.2_ori"    "PCLO.1_ori"     "PBXIP1.2_ori"   "PBXIP1.1_ori"  
#>  [21] "NEFH.1_ori"     "MYO1D.1_ori"    "MVP.1_ori"      "METAP2.1_ori"  
#>  [25] "MBP.3_ori"      "MBP.2_ori"      "MBP.1_ori"      "HP.1_ori"      
#>  [29] "HBD.2_ori"      "HBD.1_ori"      "GSTM1.2_ori"    "GFAP.2_ori"    
#>  [33] "ERMN.1_ori"     "ENO2.2_ori"     "ENO2.1_ori"     "DKK3.1_ori"    
#>  [37] "CRABP1.1p_ori"  "CRABP1.1_ori"   "BSN.2_ori"      "ALDH1A1.3_ori" 
#>  [41] "ALDH1A1.1_ori"  "ACTR3B.1_ori"   "TIMP3.2_zsc"    "TIMP3.1_zsc"   
#>  [45] "TH.2_zsc"       "SLK.1_zsc"      "SERPINA3.1_zsc" "SCG5.2_zsc"    
#>  [49] "SCG5.1_zsc"     "S100B.2_zsc"    "S100B.1_zsc"    "PTPRN.2_zsc"   
#>  [53] "PREPL.2_zsc"    "PREPL.1_zsc"    "PLIN3.1_zsc"    "PLEC.1_zsc"    
#>  [57] "PDYN.2_zsc"     "PDYN.1_zsc"     "PDE2A.2_zsc"    "PCLO.1_zsc"    
#>  [61] "PBXIP1.2_zsc"   "PBXIP1.1_zsc"   "NEFH.1_zsc"     "MYO1D.1_zsc"   
#>  [65] "MVP.1_zsc"      "METAP2.1_zsc"   "MBP.3_zsc"      "MBP.2_zsc"     
#>  [69] "MBP.1_zsc"      "HP.1_zsc"       "HBD.2_zsc"      "HBD.1_zsc"     
#>  [73] "GSTM1.2_zsc"    "GFAP.2_zsc"     "ERMN.1_zsc"     "ENO2.2_zsc"    
#>  [77] "ENO2.1_zsc"     "DKK3.1_zsc"     "CRABP1.1p_zsc"  "CRABP1.1_zsc"  
#>  [81] "BSN.2_zsc"      "ALDH1A1.3_zsc"  "ALDH1A1.1_zsc"  "ACTR3B.1_zsc"  
#>  [85] "TIMP3.2_sgn"    "TIMP3.1_sgn"    "TH.2_sgn"       "SLK.1_sgn"     
#>  [89] "SERPINA3.1_sgn" "SCG5.2_sgn"     "SCG5.1_sgn"     "S100B.2_sgn"   
#>  [93] "S100B.1_sgn"    "PTPRN.2_sgn"    "PREPL.2_sgn"    "PREPL.1_sgn"   
#>  [97] "PLIN3.1_sgn"    "PLEC.1_sgn"     "PDYN.2_sgn"     "PDYN.1_sgn"    
#> [101] "PDE2A.2_sgn"    "PCLO.1_sgn"     "PBXIP1.2_sgn"   "PBXIP1.1_sgn"  
#> [105] "NEFH.1_sgn"     "MYO1D.1_sgn"    "MVP.1_sgn"      "METAP2.1_sgn"  
#> [109] "MBP.3_sgn"      "MBP.2_sgn"      "MBP.1_sgn"      "HP.1_sgn"      
#> [113] "HBD.2_sgn"      "HBD.1_sgn"      "GSTM1.2_sgn"    "GFAP.2_sgn"    
#> [117] "ERMN.1_sgn"     "ENO2.2_sgn"     "ENO2.1_sgn"     "DKK3.1_sgn"    
#> [121] "CRABP1.1p_sgn"  "CRABP1.1_sgn"   "BSN.2_sgn"      "ALDH1A1.3_sgn" 
#> [125] "ALDH1A1.1_sgn"  "ACTR3B.1_sgn"