Nested Linear Model Analysis

A function for analysis of sigfinicance of factors (akin ANOVA) or continuous variable (regression) using nested linear models. Shuffling-based FDR estimation is available as an option.

eset_lm(eset, form.alt, form.nul, facs = NULL, norm.coef = NULL, N = NULL)

Arguments

eset

eset (or most likely eset subclass) object

form.alt

character formulation of alternative model

form.nul

character formulation of NULL model

facs

data frame with the factors in its columns. If NULL, then pData(eset) will be used.

norm.coef

vector with sample-to-sample normalization coefficients in log2 scale.

N

number of shuffles for FDR estimation. Default is NULL, that is no shuffling-based FDR estimation. Warning! Will be N-times slower. This is critical for Windows, since paralellization is not impletemted. In case of enabling shuffling-based FDR estimation, the recommended value is at least N=1000.

Value

data.frame

effect

either max contrast in case of factor or slope in case of continuous variable

F.stat

F statistic

p.value

p-value

Examples

data(srm_msnset)
head(varLabels(msnset))
#> [1] "sample.id"    "SampleNum"    "match.group"  "subject.type" "projid"      
#> [6] "age_death"   
out <- eset_lm(msnset, "~ subject.type", "~ 1")
head(out)
#>            effect.subject.typecontrol.1 effect.subject.typecontrol.2
#> TIMP3.2                   -0.2921316585                  -0.40978852
#> TIMP3.1                   -0.3266854901                  -0.59176962
#> TH.2                       0.1092869774                   0.03917179
#> SLK.1                      0.0000459538                   0.23686945
#> SERPINA3.1                 0.1878813994                   0.01644818
#> SCG5.2                     0.4475184426                   0.34126467
#>            SE.subject.typecontrol.1 SE.subject.typecontrol.2    F.stat
#> TIMP3.2                  0.44734087               0.44734087 0.4449312
#> TIMP3.1                  0.49776949               0.49776949 0.7092260
#> TH.2                     0.08631219               0.08631219 0.8230301
#> SLK.1                    0.46030084               0.46030084 0.1765058
#> SERPINA3.1               0.26035792               0.26035792 0.3194321
#> SCG5.2                   0.27817885               0.27817885 1.4129813
#>              p.value       fdr
#> TIMP3.2    0.6440808 0.8995150
#> TIMP3.1    0.4982601 0.8995150
#> TH.2       0.4465807 0.8624085
#> SLK.1      0.8388597 0.8995150
#> SERPINA3.1 0.7284444 0.8995150
#> SCG5.2     0.2556018 0.7143768
# now with shuffling
out <- eset_lm(msnset, "~ subject.type", "~ 1", N = 100)
head(out)
#>            effect.subject.typecontrol.1 effect.subject.typecontrol.2
#> TIMP3.2                   -0.2921316585                  -0.40978852
#> TIMP3.1                   -0.3266854901                  -0.59176962
#> TH.2                       0.1092869774                   0.03917179
#> SLK.1                      0.0000459538                   0.23686945
#> SERPINA3.1                 0.1878813994                   0.01644818
#> SCG5.2                     0.4475184426                   0.34126467
#>            SE.subject.typecontrol.1 SE.subject.typecontrol.2    F.stat
#> TIMP3.2                  0.44734087               0.44734087 0.4449312
#> TIMP3.1                  0.49776949               0.49776949 0.7092260
#> TH.2                     0.08631219               0.08631219 0.8230301
#> SLK.1                    0.46030084               0.46030084 0.1765058
#> SERPINA3.1               0.26035792               0.26035792 0.3194321
#> SCG5.2                   0.27817885               0.27817885 1.4129813
#>              p.value       fdr    fdr.sh
#> TIMP3.2    0.6440808 0.8995150 0.9967857
#> TIMP3.1    0.4982601 0.8995150 1.0000000
#> TH.2       0.4465807 0.8624085 0.9895238
#> SLK.1      0.8388597 0.8995150 0.9963158
#> SERPINA3.1 0.7284444 0.8995150 0.9783871
#> SCG5.2     0.2556018 0.7143768 0.9806667
head(out[order(out$p.value),])
#>          effect.subject.typecontrol.1 effect.subject.typecontrol.2
#> MBP.1                     -0.14271457                     1.980452
#> BSN.2                     -0.43567863                     1.576800
#> MBP.3                     -0.15864708                     1.918847
#> PCLO.1                    -0.19170130                     1.472722
#> ACTR3B.1                   0.01378383                     1.551503
#> MBP.2                     -0.20460049                     1.574623
#>          SE.subject.typecontrol.1 SE.subject.typecontrol.2   F.stat     p.value
#> MBP.1                   0.6454434                0.6454434 6.761440 0.003012752
#> BSN.2                   0.5946332                0.5946332 6.340877 0.004127466
#> MBP.3                   0.6589771                0.6589771 6.158593 0.004738430
#> PCLO.1                  0.5300161                0.5300161 5.904427 0.005753623
#> ACTR3B.1                0.5716558                0.5716558 4.867479 0.012967858
#> MBP.2                   0.6356316                0.6356316 4.691872 0.014932070
#>                 fdr    fdr.sh
#> MBP.1    0.06041304 0.2500000
#> BSN.2    0.06041304 0.9650000
#> MBP.3    0.06041304 0.7833333
#> PCLO.1   0.06041304 0.9925000
#> ACTR3B.1 0.10452449 1.0000000
#> MBP.2    0.10452449 0.9866667