Presence/Absence Testing

A function for testing the significance in differences of presence/absence patterns.

eset_presence_absence(
  eset,
  grouping,
  test = c("fisher", "chisq", "g", "binom"),
  ...
)

Arguments

eset

eset (or most likely eset subclass) object

grouping

character defining the column in phenoData

test

character: Fisher exact test, \(\chi^{2}\), G-test or binomial.

...

other aruments to pass to the test functions (e.g. simulate.p.value = TRUE)

Value

data.frame

p.value

p-value

effect

difference between second and first group (according to the factor levels). In case of larger number of levels, it is the absolute value of the max diffence

...

proportions of present for each group. Column names are level names.

Details

The recommended way of using the test is two-group comparison. All except "binom" options will work for larger number of groups.

Note

see also http://www.ncbi.nlm.nih.gov/pubmed/20831241

Examples

library("MSnbase")
Nsam = 20
Npep = 5
set.seed(99)
M <- matrix(rbinom(Nsam * Npep, 1, 0.5), ncol = Nsam)
pd <- data.frame(group = gl(2, Nsam/2))
fd <- data.frame(otherfdata = letters[1:Npep])
x0 <- MSnSet(M, fd, pd)

eset_presence_absence(x0, 'group', test = 'fisher')
#>     p.value effect  X1  X2
#> 1 0.7431631   -0.2 0.7 0.5
#> 2 0.1729438    0.4 0.3 0.7
#> 3 1.0000000   -0.1 0.5 0.4
#> 4 0.7431631   -0.2 0.7 0.5
#> 5 1.0000000    0.0 0.4 0.4
eset_presence_absence(x0, 'group', test = 'chisq', simulate.p.value = TRUE)
#>     p.value effect  X1  X2
#> 1 0.7381309   -0.2 0.7 0.5
#> 2 0.1704148    0.4 0.3 0.7
#> 3 1.0000000   -0.1 0.5 0.4
#> 4 0.7611194   -0.2 0.7 0.5
#> 5 1.0000000    0.0 0.4 0.4
eset_presence_absence(x0, 'group', test = 'g')
#>     p.value effect  X1  X2
#> 1 0.5592937   -0.2 0.7 0.5
#> 2 0.1347139    0.4 0.3 0.7
#> 3 0.7619870   -0.1 0.5 0.4
#> 4 0.5592937   -0.2 0.7 0.5
#> 5 0.9518328    0.0 0.4 0.4
eset_presence_absence(x0, 'group', test = 'binom')
#>      p.value effect  X1  X2
#> 1 0.17851586   -0.2 0.7 0.5
#> 2 0.01059208    0.4 0.3 0.7
#> 3 0.75390625   -0.1 0.5 0.4
#> 4 0.17851586   -0.2 0.7 0.5
#> 5 1.00000000    0.0 0.4 0.4