Wrapper for LIMMA

Convenience wrappers for differential analysis with LIMMA.

limma_a_b(eset, model.str, coef.str, ...)

limma_gen(eset, model.str, coef.str, ...)

Arguments

eset

eSet (or most likely eset subclass) object

model.str

character; formulation of the model (e.g. "~ a + b")

coef.str

character; coefficient of interest. One of colnames(pData(eset)). E.g. "a". For limma_a_b, this must be a character or factor with 2 levels or a numeric vector. For limma_gen, this can be a character or factor with at least 2 levels or a numeric vector.

...

arguments passed to lmFit

Value

data.frame. Basically output of the topTable function.

Details

limma_a_b is for either 2-factor ANOVA or linear regression. limma_gen is a more generic version and will return results for all the coefficients that match coef.str pattern. It also works the same as limma_a_b for linear regression.

If performing linear regression, the logFC column contains the slopes of the regression lines and the AveExpr column contains the y-intercepts.

limma_a_b outputs moderated t-statistics and associated p-values. limma_gen outputs moderated F-statistics and associated p-values for a set of contrasts. Also, the columns that begin with the coef.str report the log2 fold-changes. Contrasts are always relative to a reference level (if coef.str is a factor) or the first unique entry of coef.str in terms of alphanumeric order.

limma_contrasts is another LIMMA wrapper for testing custom contrasts. It has a separate documentation page.

Functions

limma_gen(): More generalized approach. The tested variable can have more than 2 groups, unlike limma_a_b.

See also

Examples

library("MSnSet.utils")
data("cptac_oca")

# Filter to features detected in at least 30 samples
ee <- oca.set
ee <- ee[rowSums(!is.na(exprs(ee))) >= 30, ]

# Note: only two levels are allowed in this case
res1 <- limma_a_b(ee,
                  model.str = "~ PLATINUM.STATUS + AGE",
                  coef.str = "PLATINUM.STATUS")
head(res1)
#>                   logFC     AveExpr          t    P.Value adj.P.Val         B
#> NP_000005.2 -0.10134987 -0.01665020 -0.4030677 0.68853940 0.9882465 -4.786651
#> NP_000007.1  0.48105900  0.05029281  2.6704526 0.01006757 0.9711979 -3.648264
#> NP_000008.1  0.11380584  0.03187806  0.7212908 0.47393885 0.9778595 -4.722203
#> NP_000009.1  0.06586753  0.03595808  0.4423633 0.66004813 0.9882465 -4.780637
#> NP_000010.1  0.10604161 -0.04849702  0.5129440 0.61014773 0.9797993 -4.768449
#> NP_000012.1 -0.02331930 -0.01768271 -0.2026986 0.84031870 0.9882465 -4.772205

# Linear regression
res2 <- limma_a_b(ee, model.str = "~ AGE", coef.str = "AGE")
head(res2)
#>                    logFC       AveExpr          t   P.Value adj.P.Val         B
#> NP_000005.2  0.003758619 -1.672939e-17  0.4968769 0.6207220 0.9773117 -7.689794
#> NP_000007.1 -0.005272805 -1.611154e-17 -0.9301632 0.3552561 0.9482298 -7.379326
#> NP_000008.1  0.002761104  1.482832e-17  0.5576797 0.5787157 0.9739896 -7.657480
#> NP_000009.1  0.005113709  6.653734e-18  1.1450797 0.2557990 0.9185004 -7.156942
#> NP_000010.1  0.003634240 -4.752667e-18  0.5787531 0.5644812 0.9739896 -7.645418
#> NP_000012.1 -0.001469325 -5.880419e-18 -0.3938889 0.6950282 0.9851141 -7.657577

# Multiple levels - "Immunoreactive" is used as the reference level
res3 <- limma_gen(ee,
                  model.str = "~ SUBTYPE",
                  coef.str = "SUBTYPE")
head(res3)
#>             SUBTYPEProliferative SUBTYPEMesenchymal SUBTYPEDifferentiated
#> NP_000005.2          -0.15149043         0.05641954           -0.15057866
#> NP_000007.1          -0.24732121        -0.15789692            0.08496949
#> NP_000008.1          -0.15921805        -0.08113290           -0.20215369
#> NP_000009.1           0.05017789         0.03855355            0.06446570
#> NP_000010.1          -0.13569585        -0.47163526            0.22698229
#> NP_000012.1           0.05628690         0.15846512            0.05044680
#>                   AveExpr          F    P.Value  adj.P.Val
#> NP_000005.2 -1.672939e-17 0.38747745 0.76234228 0.85363366
#> NP_000007.1 -1.611154e-17 1.47336052 0.22881898 0.43425511
#> NP_000008.1  1.482832e-17 0.67021902 0.57302598 0.72604829
#> NP_000009.1  6.653734e-18 0.07685254 0.97230711 0.98360868
#> NP_000010.1 -4.752667e-18 4.91941054 0.00360761 0.03327845
#> NP_000012.1 -5.880419e-18 0.56802740 0.63824459 0.77256566
#>             SE.SUBTYPEProliferative SE.SUBTYPEMesenchymal
#> NP_000005.2               0.2340000             0.2407840
#> NP_000007.1               0.1722972             0.1772924
#> NP_000008.1               0.1522815             0.1566964
#> NP_000009.1               0.1398965             0.1439524
#> NP_000010.1               0.1787178             0.1838991
#> NP_000012.1               0.1220508             0.1238611
#>             SE.SUBTYPEDifferentiated
#> NP_000005.2                0.2372278
#> NP_000007.1                0.1746739
#> NP_000008.1                0.1543820
#> NP_000009.1                0.1418262
#> NP_000010.1                0.1811830
#> NP_000012.1                0.1238611