Test Custom Contrasts with LIMMA

A convenience wrapper for differential analysis with LIMMA using custom contrasts.

limma_contrasts(
  eset,
  model.str,
  coef.str,
  contrasts,
  stat = c("t", "F"),
  var.group = character(0),
  plot = FALSE,
  col.group = coef.str,
  adjust.method = "BH",
  remove_name = TRUE,
  ...
)

Arguments

eset

eSet (or most likely eSet subclass) object

model.str

character; formulation of the model (e.g. "~ 0 + a + b"). This should be a no-intercept model (it should include 0 or -1 as terms). See lm for more details.

coef.str

character; coefficient of interest. One of colnames(pData(eset)). E.g. "a".

contrasts

character; contrasts to test. All factor levels in the contrasts should begin with the name of that factor. It is recommended to use paircomp with the name argument specified to create pairwise contrasts of the form "a-b".

stat

character; the moderated test statistic to use. Either "t" or "F".

var.group

NULL or character; the column in pData(eset) indicating groups that will have different relative quality weights. The default (NULL), means that each sample will be weighted equally.

plot

logical; whether to generate diagnostic plots. The default (TRUE) generates a barplot of weights applied to each sample and a plot of the mean-variance trend using plotSA.

col.group

character; the column in pData(eset) used to color the text of the MDS plot.

adjust.method

method for p-value adjustment. Default is "BH" (Benjamini-Hochberg). See p.adjust for details.

remove_name

if TRUE (the default) coef.str will be removed from the contrasts. This simply improves readability, but it will not work properly if coef.str is a substring of any group name.

...

additional arguments passed to eBayes.

Value

data.frame. Basically the output of topTable with additional columns feature and contrast (if stat = "t"). All columns from fData are also included.

Details

An MDS plot is used to determine the appropriate value of var.group. If samples within phenotype groups tend to cluster well and have similar dispersions, the default var.group = NULL is recommended. If samples within phenotype groups tend to cluster well, but different groups are more or less dispersed, it may be beneficial to set var.group to the name of the phenotype column. If one or more samples tend to cluster poorly with samples of the same phenotype, it may be beneficial to weight by sample. That is, set var.group to the name of the column in pData(eset) that uniquely identifies each sample. If variation between samples or groups of samples is biological rather than technical, weighting is not recommended.

The plot of the mean-variance trend helps determine whether to fit a trend line to the prior variance (default is trend = FALSE). It also helps determine if the prior variance should be robustified against outlier sample variances (robust = TRUE; default is FALSE). See eBayes for more details.

p-values are adjusted across all contrasts (globally). It is recommended to adjust p-values from similar contrasts together. See the "Multiple Testing Across Contrasts" section of the LIMMA User's Guide (linked below) for more information.

References

Liu, R., Holik, A. Z., Su, S., Jansz, N., Chen, K., Leong, H. S., Blewitt, M. E., Asselin-Labat, M. L., Smyth, G. K., & Ritchie, M. E. (2015). Why weight? Modelling sample and observational level variability improves power in RNA-seq analyses. Nucleic acids research, 43(15), e97. https://doi.org/10.1093/nar/gkv412

Phipson, B., Lee, S., Majewski, I. J., Alexander, W. S., & Smyth, G. K. (2016). ROBUST HYPERPARAMETER ESTIMATION PROTECTS AGAINST HYPERVARIABLE GENES AND IMPROVES POWER TO DETECT DIFFERENTIAL EXPRESSION. The annals of applied statistics, 10(2), 946–963. https://doi.org/10.1214/16-AOAS920

Smyth G. K. (2004). Linear models and empirical bayes methods for assessing differential expression in microarray experiments. Statistical applications in genetics and molecular biology, 3, Article3. https://doi.org/10.2202/1544-6115.1027

See also

Examples

# library(MSnSet.utils)
data(cptac_oca)

# A no-intercept model is required. Testing SUBTYPE contrasts
# with AGE as covariate
model.str <- "~ 0 + SUBTYPE + AGE"
coef.str <- "SUBTYPE"

# Create contrasts
contrasts <- paircomp(pData(oca.set)[, coef.str], name = coef.str)

# Moderated t-statistics
res1 <- limma_contrasts(oca.set, model.str = model.str,
                        coef.str = coef.str,
                        contrasts = contrasts)

# Moderated F-statistics
res2 <- limma_contrasts(oca.set, model.str = model.str,
                        coef.str = coef.str,
                        contrasts = contrasts, stat = "F")

# Count number of significant (< 0.05) features
sum(res2$adj.P.Val < 0.05)
#> [1] 1059

# Robustify mean-variance trend (based on MDS plot)
res3 <- limma_contrasts(oca.set, model.str = model.str,
                        coef.str = coef.str,
                        contrasts = contrasts,
                        trend = TRUE, robust = TRUE)

# Count number of significant (< 0.05) features in each contrast
table(res3$contrast, res3$adj.P.Val < 0.05)
#>                                
#>                                 FALSE TRUE
#>   Differentiated-Immunoreactive  8044   59
#>   Mesenchymal-Immunoreactive     7904  199
#>   Proliferative-Immunoreactive   7843  260