Example of a Longitudinal Biomarker Study

Targeted selected reaction monitoring analysis. Time points A, B, C, and D increase in the corresponding order. A & B are prior to the disease diagnosis for the cases. C & D are correspondingly after the diagnosis. Time point D is after the treatment.

data("longitudinal_biomarker_study")

Format

An MSnSet with 300 features and 236 samples.

Source

xxx study

pData

A data frame consisting of 236 rows (samples) and 12 columns:

Sample

The sample identifier. This is the SubjID and TimePoint concatenated with an underscore.

isQC

TRUE if the sample is a quality control sample.

MatchID

The unique subject identifier.

Type

"Control" or "Case".

Plate

The microplate identifier. Either "01", "02", "03", or "04".

Well

The microplate well location. PlateRow concatenated with PlateCol.

PlateRow

The row of the microplate where the sample is located. Letters A-H.

PlateCol

The column of the microplate where the sample is located. Strings "02" to "12".

TimePoint

Time point: A, B, C, or D.

Days

Number of days since disease diagnosis.

SubjID

Subject ID. This is the MatchID and a letter representing the Type ("C" for Control and "D" for Case) concatenated with an underscore.

Age

Subject age in years.

fData

A data frame consisting of 300 rows (features) and 4 columns:

Organism

The organism name: "YEAST", "BOVIN", or "HUMAN".

Protein

The gene symbol and organism name concatenated with an underscore.

Peptide

The peptide sequence.

isSpike

TRUE if Organism is "YEAST" or "BOVIN".

Author

Vlad Petyuk, 2018-11-09

Examples

data("longitudinal_biomarker_study")
head(pData(longitudinal_biomarker_study))
#>          Sample  isQC MatchID    Type Plate Well PlateRow PlateCol TimePoint
#> X03_C_B X03_C_B FALSE     X03 Control    01  B04        B       04         B
#> X03_C_A X03_C_A FALSE     X03 Control    01  A07        A       07         A
#> X03_C_D X03_C_D FALSE     X03 Control    01  B11        B       11         D
#> X03_D_B X03_D_B FALSE     X03    Case    01  A11        A       11         B
#> X03_D_A X03_D_A FALSE     X03    Case    01  B07        B       07         A
#> X03_D_D X03_D_D FALSE     X03    Case    01  A12        A       12         D
#>          Days SubjID Age
#> X03_C_B  -683  X03_C  32
#> X03_C_A -1542  X03_C  32
#> X03_C_D  1085  X03_C  32
#> X03_D_B  -707  X03_D  32
#> X03_D_A -1657  X03_D  32
#> X03_D_D   909  X03_D  32
head(fData(longitudinal_biomarker_study))
#>                   Organism    Protein           Peptide isSpike
#> NTVISVFGASGDLAK      YEAST G6PD_YEAST   NTVISVFGASGDLAK    TRUE
#> TFPALFGLFR           YEAST G6PD_YEAST        TFPALFGLFR    TRUE
#> FGNQFLNASWNR         YEAST G6PD_YEAST      FGNQFLNASWNR    TRUE
#> AVAPIDTDDVLLGQYGK    YEAST G6PD_YEAST AVAPIDTDDVLLGQYGK    TRUE
#> YQDFWIPEAYEVLIR      YEAST G6PD_YEAST   YQDFWIPEAYEVLIR    TRUE
#> DFPIANGER            BOVIN CAH2_BOVIN         DFPIANGER    TRUE
head(exprs(longitudinal_biomarker_study))
#>                        X03_C_B    X03_C_A    X03_C_D    X03_D_B      X03_D_A
#> NTVISVFGASGDLAK    0.051134485 -0.2813180 -0.3002398 0.55000596  0.043293086
#> TFPALFGLFR        -0.459198644 -0.4337984 -0.4337984 0.38252070 -0.301569007
#> FGNQFLNASWNR      -0.175419758 -0.4489848 -0.3987840 0.40467699 -0.019487397
#> AVAPIDTDDVLLGQYGK -0.003927121 -0.4085157 -0.2777886 0.70203356 -0.001681681
#> YQDFWIPEAYEVLIR   -0.842878477 -0.6367291 -0.6005953 0.04997437 -0.413217043
#> DFPIANGER          0.011771418 -0.4232525 -0.2994129 0.75004923  0.139995753
#>                       X03_D_D     X04_C_B    X04_C_A     X04_C_C     X04_C_D
#> NTVISVFGASGDLAK   -0.09147299  0.18954253 -0.2142716  0.05586619 -0.07612087
#> TFPALFGLFR        -0.32239416 -0.26079940 -0.3506370  0.12490943 -0.06675920
#> FGNQFLNASWNR      -0.30015599 -0.09190551 -0.3177403  0.16791628 -0.11547722
#> AVAPIDTDDVLLGQYGK -0.21011873 -0.10093692 -0.3790719  0.25131475 -0.13744739
#> YQDFWIPEAYEVLIR   -0.49614014 -0.65779075 -0.4380528 -0.11135018 -0.28889918
#> DFPIANGER         -0.28782493  0.16424330 -0.2740411  0.13698969 -0.25079420
#>                       X04_D_B      X04_D_A      X04_D_C     X04_D_D
#> NTVISVFGASGDLAK    0.05444831 -0.072243699  0.279270530 -0.08625056
#> TFPALFGLFR        -0.12858421 -0.140862536 -0.009349968 -0.27651764
#> FGNQFLNASWNR      -0.07277859 -0.262518328  0.085183822 -0.29065412
#> AVAPIDTDDVLLGQYGK  0.18090189 -0.026575860  0.134699560 -0.13683125
#> YQDFWIPEAYEVLIR   -0.62501571 -0.276629548  0.143489658 -0.28377414
#> DFPIANGER          0.15705413 -0.005681702  0.112712998 -0.26719828
#>                        X07_C_B    X07_C_A     X07_C_D     X07_D_B     X07_D_A
#> NTVISVFGASGDLAK   -0.048942944 -0.2585233  0.09570094  0.04019231 0.113285251
#> TFPALFGLFR        -0.026334148 -0.3843015 -0.13471031  0.06914616 0.086804828
#> FGNQFLNASWNR      -0.027356662 -0.4228009  0.09367385  0.14592531 0.000000000
#> AVAPIDTDDVLLGQYGK -0.130684263 -0.2750742  0.07531672  0.07903433 0.125484301
#> YQDFWIPEAYEVLIR   -0.107723040 -0.5591394 -0.33480224 -0.33798583 0.006290468
#> DFPIANGER          0.002363526 -0.2907132  0.01738684  0.13397735 0.060634131
#>                       X07_D_D     X05_C_B    X05_C_A      X05_C_C      X05_C_D
#> NTVISVFGASGDLAK   -0.09304341  0.02868521 -0.1030293 -0.115476695  0.091559228
#> TFPALFGLFR        -0.18044799 -0.25411509 -0.4063652 -0.009349968 -0.003732719
#> FGNQFLNASWNR       0.01540258 -0.04988888 -0.1071137 -0.187125259  0.237430182
#> AVAPIDTDDVLLGQYGK  0.14131861 -0.01520705 -0.1325256 -0.134369310  0.249430721
#> YQDFWIPEAYEVLIR   -0.45178194 -0.30023875 -0.4506328 -0.378931377 -0.276629548
#> DFPIANGER          0.08534996  0.04742963 -0.1751915  0.040552003  0.254977896
#>                       X05_D_C     X05_D_A     X05_D_B    X05_D_D     X06_C_B
#> NTVISVFGASGDLAK    0.06128853 -0.06786213 -0.01704084 0.20687691 -0.13206265
#> TFPALFGLFR         0.28339349 -0.46176344  0.03862129 0.23353298 -0.02254257
#> FGNQFLNASWNR       0.10331596 -0.25018753 -0.04588695 0.15174499 -0.09741725
#> AVAPIDTDDVLLGQYGK  0.16649717 -0.02943212 -0.05074603 0.18534323  0.04413954
#> YQDFWIPEAYEVLIR   -0.15186247 -0.85803298 -0.15747607 0.08842716 -0.25239897
#> DFPIANGER          0.30557665 -0.07618968  0.03594863 0.24942143  0.09909635
#>                       X06_C_C     X06_D_B     X06_D_C     X08_C_A     X08_C_D
#> NTVISVFGASGDLAK   -0.14096904 -0.00442568 -0.07637972 -0.16774710 -0.01704084
#> TFPALFGLFR        -0.14086254 -0.08246216 -0.19958497 -0.10834972 -0.04160090
#> FGNQFLNASWNR      -0.24712122 -0.13944046 -0.13801979 -0.20635162  0.02937860
#> AVAPIDTDDVLLGQYGK -0.07828257 -0.17427353 -0.03573586 -0.09434242  0.14841301
#> YQDFWIPEAYEVLIR   -0.21289767 -0.14348280 -0.22169164 -0.24939856 -0.45869583
#> DFPIANGER         -0.14508476 -0.08418524 -0.07668810  0.00707514  0.17097701
#>                       X08_D_A       X08_D_D     X09_C_B     X09_C_A
#> NTVISVFGASGDLAK   -0.30266046 -0.0901656066 -0.19784498 -0.01257582
#> TFPALFGLFR        -0.03585692 -0.1931777326 -0.19104829 -0.20172706
#> FGNQFLNASWNR      -0.19301382 -0.2213174678 -0.20933239  0.09004139
#> AVAPIDTDDVLLGQYGK -0.03803498 -0.0005602699 -0.10997828  0.09223383
#> YQDFWIPEAYEVLIR   -0.28070784 -0.2187543520 -0.08972287 -0.49258816
#> DFPIANGER         -0.12535706 -0.1222666768 -0.05197535  0.15493283
#>                        X09_D_B      X09_D_A     X10_C_B    X10_C_A     X10_C_C
#> NTVISVFGASGDLAK    0.001227617 -0.166093269  0.02458767 -0.3676996 -0.06221155
#> TFPALFGLFR         0.044055180 -0.104336660 -0.04544297 -0.2276833 -0.16992500
#> FGNQFLNASWNR      -0.028672390  0.116469511 -0.24253395 -0.2273475 -0.03526909
#> AVAPIDTDDVLLGQYGK -0.089565276  0.040330799  0.17892356 -0.4385729  0.20443260
#> YQDFWIPEAYEVLIR   -0.057003282 -0.284797695 -0.07106132 -0.0500253 -0.21875435
#> DFPIANGER          0.012240206 -0.008531941  0.04559880 -0.3128552 -0.03492490
#>                        X10_D_B     X10_D_A     X10_D_C     X11_C_B     X11_C_A
#> NTVISVFGASGDLAK   -0.068376918  0.12571554  0.36278883  0.06058242 -0.03428591
#> TFPALFGLFR        -0.030135718 -0.06286002  0.47928939 -0.02633415  0.12490943
#> FGNQFLNASWNR      -0.009061371  0.05444778  0.31104760  0.02050045 -0.06465834
#> AVAPIDTDDVLLGQYGK  0.209286898  0.02389277  0.38603418  0.03103892 -0.07002509
#> YQDFWIPEAYEVLIR    0.078909619 -0.26244563  0.37692731  0.04590930  0.12202515
#> DFPIANGER          0.148123634  0.15280840 -0.06675215  0.05199657 -0.03105579
#>                       X11_C_C     X11_C_D     X11_D_B     X11_D_A     X11_D_C
#> NTVISVFGASGDLAK   -0.02601254  0.18975781  0.07999169 -0.08208618 -0.04285989
#> TFPALFGLFR         0.14527723  0.03316686 -0.27426166 -0.18467877  0.02219575
#> FGNQFLNASWNR      -0.10018103  0.13303879 -0.31132118 -0.14513719 -0.02472880
#> AVAPIDTDDVLLGQYGK -0.04264423  0.15698101  0.15848774 -0.03861033  0.06892132
#> YQDFWIPEAYEVLIR    0.33955480 -0.08080631 -0.17160640 -0.30127405  0.01213642
#> DFPIANGER         -0.05983747  0.15450819 -0.09021112 -0.09071442  0.15535734
#>                       X11_D_D     X12_C_B    X12_C_A     X12_D_B    X12_D_A
#> NTVISVFGASGDLAK   -0.01530280 -0.18522780 -0.4117670 -0.05556199 -0.4104611
#> TFPALFGLFR         0.23670026  0.10251491 -0.1678296 -0.09435257 -0.3340949
#> FGNQFLNASWNR       0.02811364 -0.07142204 -0.2228226 -0.12952499 -0.4667086
#> AVAPIDTDDVLLGQYGK  0.04413954 -0.04553251 -0.1257855 -0.03861033 -0.3996193
#> YQDFWIPEAYEVLIR    0.31679255 -0.16029109 -0.3454416 -0.25039801 -0.4726238
#> DFPIANGER          0.07371887  0.01317732 -0.1429954  0.10570110 -0.3029075
#>                       X13_C_B   X13_C_A     X13_D_B     X13_D_A     X16_C_B
#> NTVISVFGASGDLAK   -0.16857473 0.2949653 -0.29089807 -0.29089807 -0.01158547
#> TFPALFGLFR        -0.12451454 0.1025149 -0.08641475 -0.29240901 -0.29240901
#> FGNQFLNASWNR      -0.28750068 0.1117003 -0.08915752 -0.39030254 -0.09741725
#> AVAPIDTDDVLLGQYGK  0.27836025 0.3148515 -0.13068426 -0.09673693  0.10166324
#> YQDFWIPEAYEVLIR   -0.11225839 0.1631390 -0.52850773 -0.18683288 -0.52004615
#> DFPIANGER          0.02855255 0.1595956 -0.06675215 -0.28840212 -0.01186435
#>                       X16_C_A     X16_C_D     X16_D_B     X16_D_A     X16_D_D
#> NTVISVFGASGDLAK    0.15820202 -0.17826565 -0.36801641 -0.47690729  0.02362185
#> TFPALFGLFR        -0.14910653  0.05126867 -0.11843149 -0.12858421  0.36670625
#> FGNQFLNASWNR       0.06928283 -0.15659858 -0.16815176  0.20197045  0.04946868
#> AVAPIDTDDVLLGQYGK  0.25272616 -0.07946607 -0.05306921  0.10427162  0.31979761
#> YQDFWIPEAYEVLIR   -0.11863194  0.04834972  0.11117194 -0.16782476 -0.09151281
#> DFPIANGER          0.22374431 -0.22736244 -0.27862110 -0.07220845  0.29829126
#>                          QC.1         QC.2          QC.3        QC.4
#> NTVISVFGASGDLAK    0.20197594 -0.024513379 -0.0776746624 -0.02976728
#> TFPALFGLFR         0.09380822 -0.047367848  0.0955537630 -0.12248401
#> FGNQFLNASWNR       0.18163456  0.009004813 -0.0418960850  0.00000000
#> AVAPIDTDDVLLGQYGK  0.13673943 -0.092549134  0.0005602699 -0.08539822
#> YQDFWIPEAYEVLIR   -0.11680805 -0.480810991 -0.2070646590 -0.18778989
#> DFPIANGER          0.06561129 -0.127937453  0.1322531847 -0.05050597
#>                          QC.5         QC.6        QC.7       QC.8     X19_C_B
#> NTVISVFGASGDLAK   -0.13179362 -0.240137348 -0.04919696 0.30473818 -0.01778635
#> TFPALFGLFR        -0.01498917  0.112894056 -0.15739790 0.20471014  0.00000000
#> FGNQFLNASWNR       0.11408686  0.108112958  0.08761465 0.09004139  0.12477794
#> AVAPIDTDDVLLGQYGK  0.02939295  0.042508452  0.17694252 0.20589061  0.11827596
#> YQDFWIPEAYEVLIR   -0.00885166 -0.206094780 -0.02929160 0.23274178 -0.11135018
#> DFPIANGER          0.07192112 -0.002363526 -0.05688416 0.04238925  0.10657945
#>                      X19_C_A     X19_C_D     X19_D_B     X19_D_A     X19_D_D
#> NTVISVFGASGDLAK   0.13603015  0.18544611  0.20921498  0.27400078  0.03516924
#> TFPALFGLFR        0.08855886  0.18352960 -0.09435257  0.05665522 -0.26527288
#> FGNQFLNASWNR      0.12122304 -0.21381514 -0.44195585  0.04447232 -0.11268400
#> AVAPIDTDDVLLGQYGK 0.07691116 -0.20494314  0.05280765 -0.15231428 -0.20365214
#> YQDFWIPEAYEVLIR   0.09316248  0.04264897  0.08605365  0.33889042 -0.90920324
#> DFPIANGER         0.19430053 -0.05590106  0.36624393 -0.10284647 -0.09323353
#>                       X20_C_B      X20_C_A    X20_D_B      X20_D_A     X21_C_B
#> NTVISVFGASGDLAK   -0.18690373 -0.006643885 0.56717615  0.008080125  0.16588237
#> TFPALFGLFR        -0.20602080  0.034987300 0.01667874 -0.116409492  0.20794127
#> FGNQFLNASWNR       0.01540258  0.207568812 0.05444778 -0.162363605  0.36113556
#> AVAPIDTDDVLLGQYGK  0.13877641 -0.119076661 0.06142400 -0.512573486  0.03103892
#> YQDFWIPEAYEVLIR   -0.01648274  0.056454768 0.34088263 -0.721499238 -0.24839981
#> DFPIANGER          0.12620233  0.120560860 0.21521844 -0.126388665  0.31852254
#>                      X21_C_A      X21_C_C      X21_C_D   X21_D_B     X21_D_A
#> NTVISVFGASGDLAK   0.24611778  0.188465654  0.006125585 0.3029481  0.13357031
#> TFPALFGLFR        0.25399708 -0.003732719  0.005581032 0.2849242 -0.03394733
#> FGNQFLNASWNR      0.03568686  0.220916896  0.083966871 0.3286227  0.08274889
#> AVAPIDTDDVLLGQYGK 0.25084397  0.290321434 -0.100336172 0.0233416  0.19516420
#> YQDFWIPEAYEVLIR   0.04183273  0.243431096  0.164639493 0.2771142 -0.04654892
#> DFPIANGER         0.33622916  0.156630115 -0.682994494 0.1008606  0.22697905
#>                       X21_D_C     X21_D_D     X22_C_B     X22_C_A     X22_C_D
#> NTVISVFGASGDLAK   -0.05046772 -0.19840831  0.24280172  0.42975548  0.16522565
#> TFPALFGLFR        -0.14704111  0.37103650 -0.36737107  0.03316686 -0.40389694
#> FGNQFLNASWNR       0.09609043 -0.02472880  0.26131281  0.12359395 -0.06465834
#> AVAPIDTDDVLLGQYGK -0.02144880  0.02168682  0.03432525  0.12702427 -0.16419353
#> YQDFWIPEAYEVLIR   -0.06577349  0.35868984 -0.13515153 -0.18396565 -0.22169164
#> DFPIANGER          0.14299557  0.14085348  0.23342686  0.43359148 -0.25360922
#>                       X22_D_B      X22_D_A    X22_D_D      X14_C_B     X14_C_A
#> NTVISVFGASGDLAK    0.35359478 -0.007630849  0.0209625  0.120756266  0.48553257
#> TFPALFGLFR         0.10944263  0.155354303  0.2571198  0.053066421  0.34191957
#> FGNQFLNASWNR       0.17935722  0.103315963  0.3731414  0.012846880  0.19296765
#> AVAPIDTDDVLLGQYGK  0.21267521  0.003357818  0.1270243 -0.003365433 -0.25621515
#> YQDFWIPEAYEVLIR   -0.03272659  0.116608749  0.0939502  0.147289226 -0.02159265
#> DFPIANGER          0.41916968  0.181852699 -0.1053869  0.034103164  0.07820347
#>                       X14_C_C     X14_C_D     X14_D_B     X14_D_A     X14_D_C
#> NTVISVFGASGDLAK    0.00465794 -0.08885941  0.13245082  0.04947472 -0.06375042
#> TFPALFGLFR         0.07800251  0.16868183  0.20471014 -0.02823368  0.02769173
#> FGNQFLNASWNR       0.13539037  0.27417496  0.08761465 -0.09465875  0.20421241
#> AVAPIDTDDVLLGQYGK -0.07710003  0.06838708 -0.11786022 -0.51978497 -0.05597845
#> YQDFWIPEAYEVLIR    0.09788232  0.33822574  0.13891693  0.09709675  0.26457503
#> DFPIANGER          0.13655974  0.19140563  0.12620233  0.21032379  0.08890997
#>                      X14_D_D     X15_C_A     X15_C_D     X15_D_A     X15_D_D
#> NTVISVFGASGDLAK   0.03923688 -0.08130670  0.14405134 -0.03051940  0.18523019
#> TFPALFGLFR        0.33602500  0.06202173  0.02586206 -0.16573720  0.05306642
#> FGNQFLNASWNR      0.33681990  0.12950422  0.26131281  0.04322053  0.02684756
#> AVAPIDTDDVLLGQYGK 0.24518264 -0.20171780 -0.15169176 -0.20429749 -0.02372525
#> YQDFWIPEAYEVLIR   0.42216455  0.08921746 -0.23448942 -0.33586265  0.22413293
#> DFPIANGER         0.37720394  0.05199657  0.47431124  0.10086058 -0.04366854
#>                        X01_C_B     X01_C_A     X01_D_B     X01_D_A    X17_C_B
#> NTVISVFGASGDLAK    0.101433661  0.08045618 -0.12027236 -0.13907521 0.08856058
#> TFPALFGLFR         0.225584226 -0.22768329 -0.01687382  0.27417496 0.19497313
#> FGNQFLNASWNR       0.056930909  0.05568988  0.04572304 -0.01556880 0.10930972
#> AVAPIDTDDVLLGQYGK  0.153962815  0.17048527 -0.26023566 -0.09793568 0.19075304
#> YQDFWIPEAYEVLIR   -0.002102197  0.14804794  0.31138412  0.02458425 0.18030085
#> DFPIANGER          0.224958190  0.10921131  0.07416795 -0.18644379 0.20459231
#>                     X17_C_A     X17_D_B     X17_D_A    X18_C_B     X18_C_A
#> NTVISVFGASGDLAK   0.1110136 -0.12481629 -0.05428674  0.0031888 -0.06323728
#> TFPALFGLFR        0.3241631  0.18844509  0.04405518  0.2303587 -0.08048992
#> FGNQFLNASWNR      0.1793572  0.01667874  0.32244417  0.3089658 -0.06061528
#> AVAPIDTDDVLLGQYGK 0.1659979 -0.03516165  0.24186990 -0.2873293 -0.05306921
#> YQDFWIPEAYEVLIR   0.3749843  0.15259186  0.18326478 -0.4679662  0.07093009
#> DFPIANGER         0.1374195 -0.03589380  0.22050230 -0.0392901 -0.12587277
#>                        X18_D_B     X18_D_A    X23_C_B      X23_C_A    X23_D_B
#> NTVISVFGASGDLAK   -0.082866086 -0.20179294 -0.2145565 -0.153749572 0.01562906
#> TFPALFGLFR        -0.020650509 -0.10834972  0.1802433 -0.074589316 0.13512927
#> FGNQFLNASWNR       0.179357217 -0.11967721  0.1702117  0.161007907 0.25591955
#> AVAPIDTDDVLLGQYGK  0.006707731 -0.02201758  0.2199092 -0.081243155 0.35219913
#> YQDFWIPEAYEVLIR   -0.047417233 -0.04221520  0.5550990  0.240588346 0.45011862
#> DFPIANGER          0.197189638  0.01458186  0.1713968  0.003307081 0.18601392
#>                       X23_D_A   X24_C_B     X24_C_A     X24_C_C    X24_C_D
#> NTVISVFGASGDLAK   0.002208542 0.3319114 -0.14802513  0.12030458 0.16040058
#> TFPALFGLFR        0.376790030 0.4725885 -0.10033473  0.21277442 0.30919725
#> FGNQFLNASWNR      0.218700774 0.3347750 -0.08094481 -0.01426497 0.23852438
#> AVAPIDTDDVLLGQYGK 0.171480577 0.2232727  0.04902176 -0.21987330 0.07052283
#> YQDFWIPEAYEVLIR   0.447655628 1.5103652  0.61559349  0.11583331 0.50563694
#> DFPIANGER         0.119690981 0.2209079  0.01738684 -0.00520721 0.17558827
#>                       X24_D_B   X24_D_A    X24_D_C      X24_D_D     X25_C_B
#> NTVISVFGASGDLAK   -0.16004538 0.2275747 0.01465722 -0.062724321  0.01514322
#> TFPALFGLFR         0.54208204 0.3681511 0.26799639 -0.003732719  0.03862129
#> FGNQFLNASWNR       0.15985329 0.2341426 0.32656616 -0.181260636 -0.03791623
#> AVAPIDTDDVLLGQYGK  0.08327132 0.2432906 0.10999351  0.022790217 -0.02486482
#> YQDFWIPEAYEVLIR    0.84907101 0.4847668 0.77585605  0.110393575  0.49613377
#> DFPIANGER          0.06786797 0.3673437 0.10174188 -0.069725757  0.07551438
#>                       X25_C_D      X25_D_B    X25_D_D     X26_C_B     X26_C_A
#> NTVISVFGASGDLAK    0.01611473  0.310887027 0.25375711 -0.10302932 -0.21313271
#> TFPALFGLFR        -0.01122726  0.007436589 0.09206055  0.07976728 -0.26527288
#> FGNQFLNASWNR       0.13186156 -0.009061371 0.08761465  0.01922767 -0.24406142
#> AVAPIDTDDVLLGQYGK -0.07002509  0.152451349 0.09905013 -0.45922142 -0.16293846
#> YQDFWIPEAYEVLIR    0.44765563  0.221251861 0.65882355  0.53493720  0.04264897
#> DFPIANGER         -0.02142816  0.188090041 0.21806600 -0.02094846 -0.26378900
#>                       X26_C_C      X26_C_D    X26_D_B    X26_D_A     X26_D_C
#> NTVISVFGASGDLAK   -0.21769369  0.014900240 0.20496112 0.03708486 -0.19784498
#> TFPALFGLFR         0.11461668  0.304677052 0.26799639 0.16536148 -0.04929530
#> FGNQFLNASWNR       0.03190522  0.074194132 0.07051222 0.15174499  0.09488265
#> AVAPIDTDDVLLGQYGK -0.06591401 -0.009556059 0.17148058 0.12445674  0.04522590
#> YQDFWIPEAYEVLIR    0.63087180  0.576674504 0.39749169 0.50859389 -0.15092899
#> DFPIANGER          0.16339937 -0.080181919 0.17433212 0.13698969 -0.05540976
#>                       X26_D_D      QC.9        QC.10       QC.11       QC.12
#> NTVISVFGASGDLAK   -0.04691241 0.4489482  0.124364697  0.16959812 -0.32647692
#> TFPALFGLFR         0.01852009 0.2477313 -0.459198644 -0.16156153 -0.09634387
#> FGNQFLNASWNR      -0.07413642 0.2731075  0.217591435  0.14592531 -0.09603734
#> AVAPIDTDDVLLGQYGK  0.06624817 0.2971773  0.065712949  0.22471178 -0.17997464
#> YQDFWIPEAYEVLIR   -0.00126073 0.1442504 -0.087042108  0.02458425 -0.01733314
#> DFPIANGER          0.18143592 0.2338289  0.005192341  0.17600675 -0.23346264
#>                         QC.13        QC.14        QC.15       QC.16    X02_C_B
#> NTVISVFGASGDLAK   -0.49825035 -0.041595811 -0.029516661  0.11010400 0.36240691
#> TFPALFGLFR        -0.14910653  0.114616678  0.007436589  0.43585597 0.05306642
#> FGNQFLNASWNR      -0.01556880  0.118848233  0.179357217  0.23633516 0.05320462
#> AVAPIDTDDVLLGQYGK -0.09973568  0.005033747 -0.328948414 -0.19400627 0.35483019
#> YQDFWIPEAYEVLIR   -0.00126073  0.303914281  0.553954427  0.89815169 0.21257397
#> DFPIANGER         -0.22294208  0.133115525 -0.011864351  0.03871242 0.04742963
#>                       X02_C_A    X02_C_D    X02_D_B     X02_D_A     X02_D_D
#> NTVISVFGASGDLAK   -0.08729353  0.1050907 -0.2046195 -0.01902974 -0.09251975
#> TFPALFGLFR        -0.13880888 -0.1889220 -0.1532463 -0.21032736 -0.16364786
#> FGNQFLNASWNR      -0.16236361 -0.1154772 -0.1168759  0.02304265 -0.14228601
#> AVAPIDTDDVLLGQYGK  0.06998919 -0.1535601 -0.3184305 -0.22903745 -0.07473786
#> YQDFWIPEAYEVLIR    0.24343110 -0.1916243 -0.2474018 -0.23448942 -0.14441148
#> DFPIANGER         -0.24518059 -0.1212380 -0.3527424 -0.22625608 -0.21414178
#>                       X27_C_B     X27_C_A     X27_C_C     X27_D_B      X27_D_A
#> NTVISVFGASGDLAK   -0.07845219  0.09983079 -0.22427547  0.17068917  0.012711574
#> TFPALFGLFR        -0.01687382  0.03316686 -0.23863775 -0.13471031 -0.128584206
#> FGNQFLNASWNR      -0.11547722 -0.21681139 -0.22734755  0.00000000 -0.198926511
#> AVAPIDTDDVLLGQYGK -0.23627892  0.01781826 -0.28118874 -0.22183215 -0.116644810
#> YQDFWIPEAYEVLIR   -0.16217082  0.20384356 -0.05612919  0.08446914 -0.007161331
#> DFPIANGER         -0.14874842 -0.07618968 -0.44368035 -0.13466814 -0.136225849
#>                       X27_D_C     X28_C_B      X28_C_A      X28_C_D     X28_D_B
#> NTVISVFGASGDLAK    0.28109020  0.02796296  0.099143305 -0.067862129  0.12526540
#> TFPALFGLFR        -0.05509314  0.08504866 -0.009349968 -0.060914380 -0.08246216
#> FGNQFLNASWNR      -0.17250816 -0.17687776 -0.114079934 -0.019487397 -0.16525478
#> AVAPIDTDDVLLGQYGK -0.06122993 -0.26090684 -0.047847302  0.009493395 -0.14300453
#> YQDFWIPEAYEVLIR    0.10415151  0.20675957 -0.046548924 -0.151862468  0.12819052
#> DFPIANGER         -0.15242140 -0.27862110 -0.164556044 -0.183219888 -0.17946775
#>                       X28_D_A      X28_D_D     X30_C_B    X30_C_A     X30_C_C
#> NTVISVFGASGDLAK   -0.05021348 -0.001227617  0.01125062 0.05633851 -0.13448622
#> TFPALFGLFR        -0.28785073 -0.009349968 -0.08443710 0.11116938 -0.04352066
#> FGNQFLNASWNR      -0.13660051 -0.105724534 -0.05122532 0.01922767 -0.12248401
#> AVAPIDTDDVLLGQYGK -0.03401392 -0.097935681  0.05766068 0.02829459 -0.11725239
#> YQDFWIPEAYEVLIR    0.27919353 -0.064015179 -0.13422879 0.01380235 -0.28889918
#> DFPIANGER         -0.27232736 -0.266060957 -0.30524198 0.03317954 -0.09373788
#>                       X30_D_B     X30_D_A     X30_D_C      X29_C_B    X29_C_A
#> NTVISVFGASGDLAK    0.10828299 -0.05403182  0.02047846  0.200694684 -0.1777101
#> TFPALFGLFR        -0.29012807 -0.24525057 -0.02443711  0.007436589  0.0349873
#> FGNQFLNASWNR       0.08396687  0.17021173 -0.15516092 -0.152289883 -0.2183119
#> AVAPIDTDDVLLGQYGK  0.08221324 -0.06181461 -0.05831008 -0.112398910 -0.3107659
#> YQDFWIPEAYEVLIR   -0.26143781 -0.32108730 -0.29817036 -0.143482803 -0.4255815
#> DFPIANGER         -0.19399436 -0.14665376 -0.10895097 -0.156103760 -0.2575595
#>                       X29_C_C     X29_C_D     X29_D_B     X29_D_A     X29_D_C
#> NTVISVFGASGDLAK   -0.42621053  0.23092590  0.05397537 -0.09461554  0.86645676
#> TFPALFGLFR        -0.02443711  0.18024328 -0.12858421 -0.13675814 -0.13675814
#> FGNQFLNASWNR      -0.05256301 -0.20486354  0.01284688 -0.12107992 -0.02735666
#> AVAPIDTDDVLLGQYGK -0.44009233  0.01615712  0.07956463 -0.11786022 -0.47082182
#> YQDFWIPEAYEVLIR    0.09788232  0.07412720  0.00126073  0.12511113  0.26597366
#> DFPIANGER         -0.27690190  0.09467633 -0.06427882 -0.26606096  0.05563974
#>                       X29_D_D     X31_C_A     X31_C_C     X31_D_A     X31_D_C
#> NTVISVFGASGDLAK   -0.05683837 -0.24245913  0.32604752 -0.06169895  0.03060945
#> TFPALFGLFR         0.82377042 -0.13675814  0.23670026  0.02769173  0.18024328
#> FGNQFLNASWNR      -0.22131747 -0.13801979  0.03820245 -0.07821760  0.08396687
#> AVAPIDTDDVLLGQYGK  0.15496958 -0.03976172  0.04848011  0.06464191 -0.11968526
#> YQDFWIPEAYEVLIR   -0.38002474 -0.07371252 -0.02159265 -0.15092899 -0.09779516
#> DFPIANGER         -0.01855235 -0.15821219  0.06831888 -0.10793176  0.03410316
#>                       X32_C_A     X32_C_C     X32_D_A    X32_D_C     X38_C_B
#> NTVISVFGASGDLAK   -0.02926608 -0.32217312  0.15732166 0.13446529 -0.18578623
#> TFPALFGLFR        -0.06480830 -0.06480830 -0.01687382 0.21277442  0.33010624
#> FGNQFLNASWNR      -0.21531248 -0.36183288  0.09004139 0.08396687  0.25483848
#> AVAPIDTDDVLLGQYGK -0.03057621 -0.21271351  0.04033080 0.05927473 -0.07710003
#> YQDFWIPEAYEVLIR   -0.07991767 -0.04481387 -0.05700328 0.51331244  0.14272854
#> DFPIANGER         -0.21249768 -0.20812256 -0.05639252 0.07416795 -0.02575264
#>                        X38_C_A     X38_C_C     X38_C_D     X38_D_B     X38_D_A
#> NTVISVFGASGDLAK    0.170907283  0.09363157 -0.05965040  0.11146825  0.07813225
#> TFPALFGLFR        -0.047367848  0.68805599 -0.16156153  0.09903854  0.03862129
#> FGNQFLNASWNR      -0.079580561 -0.28435413 -0.19301382  0.06682089 -0.03924163
#> AVAPIDTDDVLLGQYGK  0.169489279  0.28389312 -0.08183600  0.10687529  0.06249743
#> YQDFWIPEAYEVLIR    0.320835599  0.08605365  0.03774462 -0.26143781 -0.06138172
#> DFPIANGER         -0.009007528 -0.12123802 -0.28494242  0.15408343  0.01691973
#>                       X38_D_C     X38_D_D      X33_C_A     X33_C_C    X33_C_D
#> NTVISVFGASGDLAK    0.37798323 -0.08078727 -0.003194817  0.02120446  0.1178178
#> TFPALFGLFR         0.12148667 -0.04160090  0.092060553  0.07976728 -0.1678296
#> FGNQFLNASWNR      -0.08778548  0.02177211  0.082748893 -0.01296231  0.3388619
#> AVAPIDTDDVLLGQYGK  0.05981235  0.05550579 -0.016906667  0.03869540 -0.1831517
#> YQDFWIPEAYEVLIR    0.05240793  0.18843706  0.232026332  0.28265241  0.2652745
#> DFPIANGER         -0.12897090 -0.13622585 -0.231241403 -0.16561607 -0.2196357
#>                        X33_D_A    X33_D_C     X33_D_D    X34_C_B     X34_C_A
#> NTVISVFGASGDLAK   -0.787602837 0.44119325 -0.02451338 0.14471776  0.11464650
#> TFPALFGLFR        -0.350637050 0.28950662  0.19660054 0.41783071  0.62297876
#> FGNQFLNASWNR       0.219809260 0.29009305 -0.34370626 0.07419413 -0.12670447
#> AVAPIDTDDVLLGQYGK  0.005033747 0.11724325 -0.26628753 0.27558583  0.01560299
#> YQDFWIPEAYEVLIR   -0.157476075 0.04428005  0.46237088 0.29160686  0.11039357
#> DFPIANGER         -0.065267640 0.14470695  0.02762537 0.08178114 -0.36669332
#>                       X34_D_B     X34_D_A    X35_C_A     X35_C_D     X35_D_A
#> NTVISVFGASGDLAK    0.13893184 -0.44280104 -0.3428872 -0.09855339  0.26135621
#> TFPALFGLFR         0.34485785  0.22876898  0.3061854 -0.31078754 -0.22115032
#> FGNQFLNASWNR      -0.02472880 -0.38692386  0.2286466  0.18163456 -0.33391418
#> AVAPIDTDDVLLGQYGK -0.03803498  0.08802311 -0.1136107  0.37484377  0.06571295
#> YQDFWIPEAYEVLIR    0.61504484  0.32083560  0.1653891 -0.41994814 -0.09420190
#> DFPIANGER         -0.13414928 -0.03395665 -0.1698640 -0.34852312  0.18060199
#>                       X35_D_D   X36_C_B    X36_C_C    X36_C_D     X36_D_B
#> NTVISVFGASGDLAK    0.22862278 0.1396006 0.01076330 0.12818879  0.42372788
#> TFPALFGLFR         0.70530644 0.1025149 0.11116938 0.44409996  0.24616059
#> FGNQFLNASWNR      -0.25018753 0.2773727 0.15290611 0.09246405 -0.05256301
#> AVAPIDTDDVLLGQYGK  0.04848011 0.1079154 0.07956463 0.02554503 -0.16105791
#> YQDFWIPEAYEVLIR    0.46419977 0.5191890 0.12202515 0.62325287  0.62543382
#> DFPIANGER         -0.07171157 0.2160326 0.19430053 0.12923093 -0.00805651
#>                      X36_D_C    X36_D_D      X37_C_B   X37_C_A     X37_D_B
#> NTVISVFGASGDLAK    0.2566372 -0.3160470 -0.044631467 0.1384858 -0.01505468
#> TFPALFGLFR         0.0349873  0.1163372  0.191712803 0.6845810  0.01114056
#> FGNQFLNASWNR      -0.1015649  0.1247779 -0.006466596 0.2164812 -0.19153942
#> AVAPIDTDDVLLGQYGK -0.1337545 -0.2880132  0.159491352 0.2751229 -0.12823278
#> YQDFWIPEAYEVLIR    0.2154724  0.4209095  0.438381790 0.9936831  0.39557618
#> DFPIANGER          0.2890358 -0.2015847 -0.069229730 0.1536585  0.09068669
#>                       X37_D_A       QC.17      QC.18        QC.19       QC.20
#> NTVISVFGASGDLAK    0.07696888 -0.04539138 -0.2729873 -0.088076258 -0.01803494
#> TFPALFGLFR         0.10251491  0.07623559 -0.1491065 -0.086414752  0.05665522
#> FGNQFLNASWNR      -0.04455543  0.03820245 -0.1754198 -0.239483857 -0.06061528
#> AVAPIDTDDVLLGQYGK  0.18189003  0.04413954 -0.2251028 -0.005051153 -0.15356013
#> YQDFWIPEAYEVLIR    0.26737095 -0.04481387 -0.2012551 -0.241427853 -0.02073974
#> DFPIANGER         -0.19886925 -0.07718670 -0.2649245 -0.115593462 -0.15610376
#>                        QC.21       QC.22       QC.23     QC.24
#> NTVISVFGASGDLAK   -0.2723941 -0.28251202  0.20133545 0.3260475
#> TFPALFGLFR         0.0000000  0.13173070  0.16203346 0.3681511
#> FGNQFLNASWNR      -0.2318867 -0.04988888  0.26453912 0.1861785
#> AVAPIDTDDVLLGQYGK -0.0118138 -0.13929739 -0.02943212 0.1468957
#> YQDFWIPEAYEVLIR    0.3691395  0.38080552  0.26177368 0.6956156
#> DFPIANGER         -0.2301321 -0.24686238 -0.12535706 0.1591723