CPTAC Ovarian Cancer Proteomics Dataset

Processed iTRAQ data from both PNNL and JHU sites.

data("cptac_oca")

Format

An MSnSet with 8103 features and 73 samples.

Source

CPTAC study

pData

A data frame consisting of 73 rows (samples) and 39 columns:

Tumor.Nuclei

Percent of cells in the tissue sample with tumor nuclei.

Necrosis

Percent of tumor nuclei that are necrotic.

PNNL.Weight..mg.

Weight of the non-tumor tissue in milligrams.

Residual.Tumor.Tissue..mg.

Weight of the tumor tissue in milligrams.

TSS

primary_therapy_outcome_success

One of "COMPLETE.RESPONSE", "STABLE.DISEASE", or "PROGRESSIVE.DISEASE".

days_to_death

Days between

Platinum.Refractory

tumor_grade

tumor_residual_disease

tumor_stage

Tumor stage.

anatomic_organ_subdivision

Either "left", "right", or "bilateral".

days_to_birth

Negative age (in days) of subject.

days_to_tumor_recurrence

Days before first tumor recurrence.

race

The race of the subject.

site_of_tumor_first_recurrence

"LOCO.REGIONAL" or "METASTASIS".

Batch

The sample batch.

additional_chemo_therapy

"YES" if the subject received additional chemotherapy; "NO" otherwise.

additional_drug_therapy

"YES" if the subject received additional drug therapy; "NO" otherwise.

additional_pharmaceutical_therapy

"YES" if the subject received additional pharmaceutical therapy; "NO" otherwise.

additional_radiation_therapy

"YES" if the subject received additional radiation therapy; "NO" otherwise.

targeted_molecular_therapy

"YES" if the subject received targeted molecular therapy; "NO" otherwise.

year_of_initial_pathologic_diagnosis

The year when subjects were first diagnosed with ovarian cancer.

iTRAQ_Batch

The iTRAQ batch identifier.

iTRAQ_ID

ReporterIon

SUBTYPE

Ovarian cancer subtype. Either "Differentiated", "Immunoreactive", "Mesenchymal", or "Proliferative".

AGE

Subject age in years.

PLATINUM.STATUS

The tumor's response to platinum therapy. Either "SENSITIVE" or "RESISTANT".

TUMORRESIDUALDISEASE

SURVIVALSTATUS

The survival status of the patient. Either "LIVING" or "DECEASED".

SURVIVALMONTHS

RECURRENCE.MONTHS

Differentiated.binary

TRUE if SUBTYPE is "Differentiated"; FALSE otherwise.

Immunoreactive.binary

TRUE if SUBTYPE is "Immunoreactive"; FALSE otherwise.

Mesenchymal.binary

TRUE if SUBTYPE is "Mesenchymal"; FALSE otherwise.

Proliferative.binary

TRUE if SUBTYPE is "Proliferative"; FALSE otherwise.

SILHOUETTE.WIDTH

fake

fData

RefSeq

The NCBI reference sequence.

Examples

data("cptac_oca")
head(pData(oca.set))
#>              Tumor.Nuclei Necrosis PNNL.Weight..mg. Residual.Tumor.Tissue..mg.
#> TCGA-13-1484           95        0               57                         90
#> TCGA-13-1488           85        0               78                        460
#> TCGA-13-1489           95        5               77                         98
#> TCGA-13-1494           80        1               75                        221
#> TCGA-13-1495           70        5               83                        165
#> TCGA-13-1499           60        1               75                        150
#>              TSS primary_therapy_outcome_success days_to_death
#> TCGA-13-1484 MSK               COMPLETE.RESPONSE            NA
#> TCGA-13-1488 MSK               COMPLETE.RESPONSE          2154
#> TCGA-13-1489 MSK               COMPLETE.RESPONSE            NA
#> TCGA-13-1494 MSK               COMPLETE.RESPONSE            NA
#> TCGA-13-1495 MSK               COMPLETE.RESPONSE            NA
#> TCGA-13-1499 MSK               COMPLETE.RESPONSE            NA
#>              Platinum.Refractory tumor_grade tumor_residual_disease tumor_stage
#> TCGA-13-1484                  X0          G3                   <NA>        IIIC
#> TCGA-13-1488                  X0          G3              X11.20.mm          IV
#> TCGA-13-1489                  X0          G2                   <NA>        IIIC
#> TCGA-13-1494                  X0          G3                   <NA>          IV
#> TCGA-13-1495                  X0          G2               X1.10.mm        IIIC
#> TCGA-13-1499                  X0          G3                   <NA>        IIIC
#>              anatomic_organ_subdivision days_to_birth days_to_tumor_recurrence
#> TCGA-13-1484                       <NA>        -22698                      496
#> TCGA-13-1488                  Bilateral        -21726                      351
#> TCGA-13-1489                  Bilateral        -25768                      806
#> TCGA-13-1494                  Bilateral        -15993                      693
#> TCGA-13-1495                  Bilateral        -22046                      977
#> TCGA-13-1499                       <NA>        -20668                      582
#>               race site_of_tumor_first_recurrence Batch
#> TCGA-13-1484 WHITE                     METASTASIS   X14
#> TCGA-13-1488 WHITE                     METASTASIS   X14
#> TCGA-13-1489 WHITE                  LOCO.REGIONAL   X14
#> TCGA-13-1494 WHITE                     METASTASIS   X14
#> TCGA-13-1495 WHITE                     METASTASIS   X14
#> TCGA-13-1499 WHITE                     METASTASIS   X14
#>              additional_chemo_therapy additional_drug_therapy
#> TCGA-13-1484                       NO                      NO
#> TCGA-13-1488                      YES                      NO
#> TCGA-13-1489                      YES                      NO
#> TCGA-13-1494                       NO                      NO
#> TCGA-13-1495                     <NA>                    <NA>
#> TCGA-13-1499                      YES                      NO
#>              additional_pharmaceutical_therapy additional_radiation_therapy
#> TCGA-13-1484                               YES                           NO
#> TCGA-13-1488                               YES                           NO
#> TCGA-13-1489                                NO                           NO
#> TCGA-13-1494                                NO                           NO
#> TCGA-13-1495                              <NA>                         <NA>
#> TCGA-13-1499                                NO                           NO
#>              targeted_molecular_therapy year_of_initial_pathologic_diagnosis
#> TCGA-13-1484                         NO                                 2000
#> TCGA-13-1488                         NO                                 2002
#> TCGA-13-1489                         NO                                 2002
#> TCGA-13-1494                         NO                                 2004
#> TCGA-13-1495                        YES                                 2004
#> TCGA-13-1499                         NO                                 2005
#>              iTRAQ_Batch iTRAQ_ID ReporterIon        SUBTYPE AGE
#> TCGA-13-1484          B3     B3S2     Ion_116 Immunoreactive  62
#> TCGA-13-1488          B5     B5S5     Ion_115 Immunoreactive  59
#> TCGA-13-1489          B2     B2S5     Ion_114  Proliferative  70
#> TCGA-13-1494          B1     B1S4     Ion_115    Mesenchymal  43
#> TCGA-13-1495          B5     B5S1     Ion_116 Differentiated  60
#> TCGA-13-1499          B4     B4S5     Ion_115    Mesenchymal  56
#>              PLATINUM.STATUS TUMORRESIDUALDISEASE SURVIVALSTATUS SURVIVALMONTHS
#> TCGA-13-1484       SENSITIVE                 <NA>         LIVING           60.0
#> TCGA-13-1488       SENSITIVE           Suboptimal         LIVING           60.0
#> TCGA-13-1489       SENSITIVE                 <NA>         LIVING           60.0
#> TCGA-13-1494       SENSITIVE                 <NA>         LIVING           38.7
#> TCGA-13-1495       SENSITIVE              Optimal         LIVING           49.9
#> TCGA-13-1499       SENSITIVE                 <NA>         LIVING           38.9
#>              RECURRENCE.MONTHS Differentiated.binary Immunoreactive.binary
#> TCGA-13-1484          16.53333                  TRUE                  TRUE
#> TCGA-13-1488          11.70000                  TRUE                  TRUE
#> TCGA-13-1489          26.86667                  TRUE                 FALSE
#> TCGA-13-1494          23.10000                  TRUE                 FALSE
#> TCGA-13-1495          32.56667                  TRUE                 FALSE
#> TCGA-13-1499          19.40000                 FALSE                  TRUE
#>              Mesenchymal.binary Proliferative.binary SILHOUETTE.WIDTH fake
#> TCGA-13-1484              FALSE                 TRUE      0.134100414  raz
#> TCGA-13-1488              FALSE                 TRUE      0.103998708  dva
#> TCGA-13-1489              FALSE                FALSE      0.067373868  dva
#> TCGA-13-1494               TRUE                 TRUE      0.286564330  tri
#> TCGA-13-1495              FALSE                 TRUE      0.004484949  raz
#> TCGA-13-1499               TRUE                 TRUE      0.194579401  tri
head(fData(oca.set))
#>                  RefSeq
#> NP_000005.2 NP_000005.2
#> NP_000007.1 NP_000007.1
#> NP_000008.1 NP_000008.1
#> NP_000009.1 NP_000009.1
#> NP_000010.1 NP_000010.1
#> NP_000012.1 NP_000012.1
head(exprs(oca.set))
#>              TCGA-13-1484 TCGA-13-1488 TCGA-13-1489 TCGA-13-1494 TCGA-13-1495
#> NP_000005.2 -0.5779087991   0.59571791   -0.3186149   0.36099901   0.22718635
#> NP_000007.1  0.4267269746   0.06111684    0.2972885  -0.37620487   0.20516352
#> NP_000008.1 -0.0006236868   0.56155377    0.1898091  -0.19370868  -0.56384374
#> NP_000009.1 -0.1634750413   0.20441297    1.0784781  -0.27291805  -0.40425052
#> NP_000010.1  0.2640625821   0.33190721    1.0588405  -1.47655329  -0.41037443
#> NP_000012.1 -0.3144801780  -0.79552497    0.1958117   0.05471775  -0.08518555
#>             TCGA-13-1499 TCGA-23-1123 TCGA-23-1124 TCGA-24-1103 TCGA-24-1104
#> NP_000005.2   -1.1220141   -1.1300410   0.09560034   -0.5204096   -0.2135431
#> NP_000007.1   -0.2708751    0.8975027   0.98007965   -1.2594690    0.8216421
#> NP_000008.1    0.1546671    0.6939549  -0.70365329   -0.7000870    0.5289344
#> NP_000009.1   -0.1243639   -0.3872904  -0.39833921   -0.8302358    0.2628167
#> NP_000010.1   -0.3785585    1.0778733   1.11710743    0.9381929    0.9223453
#> NP_000012.1           NA   -0.1142843           NA   -0.1076079   -0.2804968
#>             TCGA-24-1105 TCGA-24-1428 TCGA-24-1430 TCGA-24-1435 TCGA-24-1466
#> NP_000005.2  -0.30059667  -1.35250860   -0.1615837   -0.5270382  0.409227342
#> NP_000007.1   0.32051344  -0.05863547    0.3887186   -0.1968494  0.001240368
#> NP_000008.1  -0.22695874   0.07966730    0.1897353   -0.3156623  0.180531910
#> NP_000009.1   0.46003817  -0.29185203    0.3666527   -0.8538317 -0.165136822
#> NP_000010.1   0.06056071   0.75747888   -0.5008875   -0.4942236  1.476288556
#> NP_000012.1   0.65312539  -0.23257462   -0.1919647           NA  0.223273912
#>             TCGA-24-1467 TCGA-24-1545 TCGA-24-1550 TCGA-24-1551 TCGA-24-1552
#> NP_000005.2    0.9269255  -0.48078871   0.40832471   2.07944884   1.30106660
#> NP_000007.1    0.9761627  -0.02941957  -0.28532322  -0.18163624   0.37716889
#> NP_000008.1   -0.5838150   0.62901696   0.14106433   0.07732310   0.66123324
#> NP_000009.1    0.4260291  -0.08624631  -0.41676996   0.33668477   0.05651396
#> NP_000010.1   -1.0459999   0.01677185  -0.84890016  -0.02450864   0.18574092
#> NP_000012.1    0.3918714   0.36547875  -0.05424123   0.21255360  -0.02118636
#>             TCGA-24-1553 TCGA-24-1555 TCGA-24-1556 TCGA-24-1562 TCGA-24-1563
#> NP_000005.2   -0.5718184   -0.2231776   0.58252246   -0.7873529    0.3780294
#> NP_000007.1   -0.2619173    0.5516532  -0.31245555    0.2759467   -0.2489930
#> NP_000008.1   -0.4120580    0.7509298  -0.29955673    0.2673165    0.4181271
#> NP_000009.1   -0.1794602   -0.2574808   0.04731945   -0.2307582    0.6428184
#> NP_000010.1   -0.1616220    0.7798359  -0.25445528   -0.6444925   -0.3077994
#> NP_000012.1   -0.1733225    0.3416447           NA           NA   -0.2177818
#>             TCGA-24-1603 TCGA-24-1604 TCGA-24-1923 TCGA-24-2020 TCGA-24-2023
#> NP_000005.2   -0.5512596  -0.38043727  -0.27229830   0.02364897   0.81066333
#> NP_000007.1    0.5072847  -0.02912583   0.26109481  -0.47901830  -0.21032446
#> NP_000008.1    0.1415423  -0.58110447   0.30774662   0.54221662  -0.08766132
#> NP_000009.1    0.2185319   0.30567448  -0.27651123  -1.15500520   0.52549000
#> NP_000010.1    0.6780313   1.43414245   0.18511760   0.22010299  -0.42121937
#> NP_000012.1           NA  -0.28408317  -0.05567753   0.33088934           NA
#>             TCGA-24-2024 TCGA-24-2027 TCGA-24-2030 TCGA-24-2260 TCGA-24-2267
#> NP_000005.2   -0.9866248   0.38535521    1.1868603  -0.24475832    0.8236943
#> NP_000007.1   -1.0769424  -0.74937946   -0.8062989   0.41002441    0.5485753
#> NP_000008.1   -0.6477276   0.40249535   -0.3081335   0.67869092    0.2784725
#> NP_000009.1    0.4769962   0.08685537    0.4532222   0.12327150   -0.1191481
#> NP_000010.1   -0.6227927   0.12151647    0.3393286  -0.08386257    0.5730623
#> NP_000012.1   -0.2886454   0.35879061    0.1525916           NA           NA
#>             TCGA-24-2288 TCGA-24-2289 TCGA-24-2290 TCGA-25-1316 TCGA-25-1319
#> NP_000005.2    0.2563410   1.11620302  -0.71798973  -0.57726543   1.00585258
#> NP_000007.1    0.1092299   1.68946968   1.18104438  -0.35081179  -0.50189139
#> NP_000008.1   -0.1773726   0.78276194   0.06736241  -0.15280132   0.01978330
#> NP_000009.1    0.2561272   0.12561392   0.17420347  -0.29857736  -0.12485167
#> NP_000010.1    0.1265961   0.05051209  -0.05103551  -0.81098606  -0.04635919
#> NP_000012.1    0.1032522   0.06054251   0.22434718  -0.07198512   0.32078353
#>             TCGA-25-1320 TCGA-25-1321 TCGA-25-1623 TCGA-25-1628 TCGA-25-1630
#> NP_000005.2   -0.5432465    0.1060496   1.08812669  -1.00131012    1.2595622
#> NP_000007.1   -0.4801079    0.5267710   0.13460442  -0.36771729   -0.5231604
#> NP_000008.1   -0.5527344    0.1305915   0.37258652   0.01739664   -0.2203713
#> NP_000009.1   -0.1877103    0.7015266   0.09065858   0.14864000   -0.4641352
#> NP_000010.1   -0.7958581    0.4600016   0.15579722  -0.62425808    0.3890216
#> NP_000012.1    0.3549958   -0.4126598   0.75336450   0.11183556           NA
#>             TCGA-25-1635 TCGA-25-2400 TCGA-25-2404 TCGA-25-2409 TCGA-29-1688
#> NP_000005.2  -0.85715442  -0.89963414  0.234648111  -0.89609765 -0.182781508
#> NP_000007.1   0.35718787  -0.87509412 -0.557214929   0.41170319  0.923875566
#> NP_000008.1   0.61774997   0.15079592  0.008817077   0.36856192  0.219132816
#> NP_000009.1   0.08695481   0.69513201  0.315980813  -0.01107263 -0.068609592
#> NP_000010.1   0.18345686  -0.05991877  0.037465314   0.65580871  0.004021449
#> NP_000012.1           NA  -0.45802719 -0.305943329   0.20063426           NA
#>             TCGA-29-1693 TCGA-29-1697 TCGA-29-1762 TCGA-29-1763 TCGA-29-1768
#> NP_000005.2   -0.2914758  -1.15482144  -0.05077071   0.08391777   1.82763374
#> NP_000007.1   -0.1043374  -0.19114545  -0.06927467  -0.27298854  -0.47218928
#> NP_000008.1   -0.7190226   0.54948017   0.05528831  -0.16589072   0.04893542
#> NP_000009.1    0.3811075   0.04166062  -0.98897363   0.05619749  -0.25996367
#> NP_000010.1    0.3271406   0.14569978   0.09760316   0.46335330  -0.25616713
#> NP_000012.1    0.1006767   0.18806513  -0.21931302   0.10740035   0.19633291
#>             TCGA-29-1774 TCGA-29-1775 TCGA-29-1776 TCGA-29-1777 TCGA-29-1785
#> NP_000005.2    0.6406915  -0.04523613   -0.1672868   -0.1961709  -0.35825954
#> NP_000007.1   -0.5202675  -0.07765198   -0.5502788    0.2989916  -0.68785094
#> NP_000008.1   -0.6291773  -0.83222411   -0.4732850   -0.5155666  -0.47953155
#> NP_000009.1   -0.3416068  -0.33427910   -0.3461370    0.1124059  -0.62327920
#> NP_000010.1   -0.3594130  -0.07604773    0.1260186   -0.5854234  -0.29725467
#> NP_000012.1    0.3072294           NA   -0.3608280   -0.2729686  -0.03182304
#>             TCGA-29-2414 TCGA-30-1866  TCGA-36-1576 TCGA-36-1581 TCGA-61-1741
#> NP_000005.2   -0.3873699  -0.45660317  0.0053833660   0.86397201 -0.466885527
#> NP_000007.1   -0.1060296   0.25063530 -0.5555723321   0.32853462  0.372612038
#> NP_000008.1   -0.8850028  -0.48327109  0.3891320980   0.01165381 -0.216857367
#> NP_000009.1   -0.2684413   0.89226117 -0.0008154525   0.25843016  0.096893550
#> NP_000010.1   -0.9278122  -0.04142591 -0.3613750233  -0.62642018 -0.638241405
#> NP_000012.1   -1.0718740  -0.38500647 -0.1474873534   0.38597958 -0.003703377
#>             TCGA-61-1900 TCGA-61-1907 TCGA-61-1911 TCGA-61-1914 TCGA-61-1919
#> NP_000005.2    0.3151234    0.4503227   0.48738227  -0.29436523   0.09948525
#> NP_000007.1   -0.4021971   -0.5907621   0.33784768  -0.67975005  -0.15644439
#> NP_000008.1   -0.1781945    1.2006729   0.19918339   0.01650301  -1.21233636
#> NP_000009.1   -0.3647852    0.2931596   1.00510328   0.15605263  -0.33076871
#> NP_000010.1   -0.4632710    0.0422894  -0.78947995   0.21956133  -0.61797951
#> NP_000012.1    0.2468636           NA   0.08012231  -0.21991642   0.09948577
#>             TCGA-61-1995 TCGA-61-2008 TCGA-61-2095
#> NP_000005.2  -0.48128766   -0.6977647    1.0105858
#> NP_000007.1   0.20218123    0.3898792    0.1031350
#> NP_000008.1   0.59312372   -0.2832290    0.1069209
#> NP_000009.1  -0.30115577    0.1824099   -0.2140891
#> NP_000010.1   0.60759755    0.4946887   -1.0199643
#> NP_000012.1  -0.01323568   -0.1831388    0.2523072