MSnSet Feature Bar Chart with (Adjusted) P-Value Coloring

MSnSet Feature Bar Chart with (Adjusted) P-Value Coloring

plot_sel_feat_with_p_vals(
  top_selected,
  msnset,
  response_colname,
  feat_name_colname = NULL,
  highlight_feats = NULL,
  highlight_reason = NULL,
  alpha = 0.05,
  title = "Frequency of\nBoruta-Selected Features"
)

Arguments

top_selected

A named vector with names as top Boruta-selected features and values as the frequency at which said features were selected

msnset

A MSnSet object that represented the input to an `rf_modeling` call (with Boruta turned on). This is required so P-values can be computed

response_colname

The column name (in pData(msnset)) of the machine learning response variable

feat_name_colname

If provided, a column in fData(msnset) by which to name the features by. If NULL, then featureNames(msnset) will be used.

highlight_feats

If provided, put a star (★) next to the label of these features. Note: this uses feat_name_colnames if passed, and not msnset's native features.

highlight_reason

Short string describing why certain features would be highlighted. Only applicable if highlight_feats is not NULL.

alpha

The alpha-level that defines statistical significance (for purposes of the color scale)

title

The title of the returned plot