Helper functions for ComBat.NA

These functions perform the Bayesian adjustments to the batch effects They do this in a iterative way.

Credit goes to the original authors of the sva package.

postmean(g.hat, g.bar, n, d.star, t2)

it.sol(sdat, g.hat, d.hat, g.bar, t2, a, b, conv = 1e-04)

postvar(sum2, n, a, b)

Arguments

g.hat

additive batch effect

g.bar

additive batch effect mean

n

number of samples in batch

d.star

bayes adjusted multiplicative effect

t2

additive batch effect variance

sdat

standardized data

d.hat

multiplicative batch effect

a

shape parameter

b

rate parameter

conv

relative change threshold, determines when to stop iteration

sum2

internal variable, used in Bayesian estimation of batch effects

Functions

it.sol():

postvar():