Protein Inference

A function for inferring parsimonious peptide to protein mapping.

protein_inference(x, Accession = "Accession", PepSeq = "PepSeq")

Arguments

x

table (data.frame, data.frame or tbl_df) with one column correspoding to peptides and the other to proteins or in general case accessions

Accession

character corresponding to column name with protein accesssions

PepSeq

character corresponding to column name with peptide sequences

Value

data.frame with two additional columns

UniqueAndRazor

TRUE if raw correspond to unique or razor peptide to protein mapping

JustifiedAccession

TRUE if protein has any unique peptide that justify its presence

Examples

library(tibble)
x <- tribble(
    ~pep, ~prot,
    "AAAAK", "P1",
    "CCCCK", "P1",
    "LLLLK", "P1",
    "AAAAK", "P2",
    "DDDDK", "P2",
    "AAAAK", "P3")
protein_inference(x, Accession = "prot", PepSeq = "pep")
#>       pep   prot UniqueAndRazor JustifiedAccession
#>    <char> <char>         <lgcl>             <lgcl>
#> 1:  AAAAK     P1           TRUE               TRUE
#> 2:  CCCCK     P1           TRUE               TRUE
#> 3:  LLLLK     P1           TRUE               TRUE
#> 4:  AAAAK     P2          FALSE               TRUE
#> 5:  DDDDK     P2           TRUE               TRUE
#> 6:  AAAAK     P3          FALSE              FALSE